+++ NEWS.md
@@ -2,6 +2,6 @@
 
 * Switch to API v3.
 * Housekeeping: pkgdown, styler, usethis, gh actions
-* Added `get_movies`
-* Added `system_status`
+* Added `get_movies`, `get_history`.
+* Added `system_status`, `update`, `root_folder`, `health`.
 * Added a `NEWS.md` file to track changes to the package.